cloud_functions/requirements.txt
Pillow == 6.0.0
google-cloud-bigquery==1.18.0
google-cloud-core==1.0.3
google-cloud-storage==1.18.0
Pillow==6.0.0
torch==1.1.0
torchvision==0.3.0